Make sure we install required packages
When checking for version, especially if we either have not installed the
package before, or want to install 2 versions of the .Net framework; before
this, if 4.6.1 was installed, 3.5.SP1 will not be considered as
install_required? This makes the comment accurate, too, and reverts to the
previous code that was in providers.