Clarify behavior on superseded .NET minor version
The former implementation of the `install_required?` helper was meant
to avoid non-idempotent behavior when a more recent .NET minor version
is already present.

There was ambiguity due to the name of the helper and the value of the
log sent in case it returned `false`.

As discussed in #55, I renamed the helper and reversed the logic to
clarify the behavior. The log message is also updated.