Fix idempotency of .NET 4.7.1 updates (KB405856)
The guard of the KB405856 package on Windows Server 2008R2 was wrong.
As explained in https://support.microsoft.com/help/4054856, it is setup
as an Update on Windows Server 2012 and newer, but appears as a product
on Windows Server 2008R2.

To fix this, changing the package name to match WS2008R2 product name
would have work, but I think it's cleaner to define one package per KB
and refers to the right one according to the Windows version.
Because the installer is the same, this introduce some code duplication
but I think we can live with it :)